requirements.txt
pandas==2.3.3
praw==7.8.1
python-dotenv==1.2.1
